+++ docs/global.js
@@ -1,2 +1,2 @@
-window.VERSION = "1.0.1";
+window.VERSION = "1.0.2";
 window.URL_PREFIX = location.pathname === "/" ? "" : location.pathname;