+++ docs/CHANGELOG.md
@@ -1,9 +1,19 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 <!-- ## [Unreleased] -->
+
+## [1.0.2] - 2020-05-17
+
+### Added
+
+- Add loading state for artboard previewer.
+
+### Changed
+
+- Update the version of `@qiqi1996/qi-design-vue`.
 
 ## [1.0.1] - 2020-05-08